Move issue bundle verification tests to verification/issuance.rs, remove SupplyInfo tests as there's no SupplyInfo now
