+++ src/supply_info.rs
@@ -28,127 +28,3 @@ impl AssetSupply {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn create_test_asset(asset_desc: &[u8]) -> AssetBase {
-        use crate::keys::{IssuanceAuthorizingKey, IssuanceValidatingKey};
-
-        let isk = IssuanceAuthorizingKey::from_bytes([1u8; 32]).unwrap();
-
-        AssetBase::derive(&IssuanceValidatingKey::from(&isk), asset_desc)
-    }
-
-    fn sum<'a, T: IntoIterator<Item = &'a AssetSupply>>(supplies: T) -> Option<NoteValue> {
-        supplies
-            .into_iter()
-            .map(|supply| supply.amount)
-            .try_fold(NoteValue::from_raw(0), |sum, value| sum + value)
-    }
-
-    #[test]
-    fn test_add_supply_valid() {
-        let mut supply_info = SupplyInfo::new();
-
-        let asset1 = create_test_asset(b"Asset 1");
-        let asset2 = create_test_asset(b"Asset 2");
-
-        let supply1 = AssetSupply::new(NoteValue::from_raw(20), false, None);
-        let supply2 = AssetSupply::new(NoteValue::from_raw(30), true, None);
-        let supply3 = AssetSupply::new(NoteValue::from_raw(10), false, None);
-        let supply4 = AssetSupply::new(NoteValue::from_raw(10), true, None);
-        let supply5 = AssetSupply::new(NoteValue::from_raw(50), false, None);
-
-        assert_eq!(supply_info.assets.len(), 0);
-
-        // Add supply1
-        assert!(supply_info.add_supply(asset1, supply1).is_ok());
-        assert_eq!(supply_info.assets.len(), 1);
-        assert_eq!(
-            supply_info.assets.get(&asset1),
-            Some(&AssetSupply::new(sum([&supply1]).unwrap(), false, None))
-        );
-
-        // Add supply2
-        assert!(supply_info.add_supply(asset1, supply2).is_ok());
-        assert_eq!(supply_info.assets.len(), 1);
-        assert_eq!(
-            supply_info.assets.get(&asset1),
-            Some(&AssetSupply::new(
-                sum([&supply1, &supply2]).unwrap(),
-                true,
-                None
-            ))
-        );
-
-        // Add supply3
-        assert!(supply_info.add_supply(asset1, supply3).is_ok());
-        assert_eq!(supply_info.assets.len(), 1);
-        assert_eq!(
-            supply_info.assets.get(&asset1),
-            Some(&AssetSupply::new(
-                sum([&supply1, &supply2, &supply3]).unwrap(),
-                true,
-                None
-            ))
-        );
-
-        // Add supply4
-        assert!(supply_info.add_supply(asset1, supply4).is_ok());
-        assert_eq!(supply_info.assets.len(), 1);
-        assert_eq!(
-            supply_info.assets.get(&asset1),
-            Some(&AssetSupply::new(
-                sum([&supply1, &supply2, &supply3, &supply4]).unwrap(),
-                true,
-                None
-            ))
-        );
-
-        // Add supply5
-        assert!(supply_info.add_supply(asset2, supply5).is_ok());
-        assert_eq!(supply_info.assets.len(), 2);
-        assert_eq!(
-            supply_info.assets.get(&asset1),
-            Some(&AssetSupply::new(
-                sum([&supply1, &supply2, &supply3, &supply4]).unwrap(),
-                true,
-                None
-            ))
-        );
-        assert_eq!(
-            supply_info.assets.get(&asset2),
-            Some(&AssetSupply::new(sum([&supply5]).unwrap(), false, None))
-        );
-    }
-
-    #[test]
-    fn test_update_finalization_set() {
-        let mut supply_info = SupplyInfo::new();
-
-        let asset1 = create_test_asset(b"Asset 1");
-        let asset2 = create_test_asset(b"Asset 2");
-        let asset3 = create_test_asset(b"Asset 3");
-
-        let supply1 = AssetSupply::new(NoteValue::from_raw(10), false, None);
-        let supply2 = AssetSupply::new(NoteValue::from_raw(20), true, None);
-        let supply3 = AssetSupply::new(NoteValue::from_raw(40), false, None);
-        let supply4 = AssetSupply::new(NoteValue::from_raw(50), true, None);
-
-        assert!(supply_info.add_supply(asset1, supply1).is_ok());
-        assert!(supply_info.add_supply(asset1, supply2).is_ok());
-        assert!(supply_info.add_supply(asset2, supply3).is_ok());
-        assert!(supply_info.add_supply(asset3, supply4).is_ok());
-
-        let mut finalization_set = HashSet::new();
-
-        supply_info.update_finalization_set(&mut finalization_set);
-
-        assert_eq!(finalization_set.len(), 2);
-
-        assert!(finalization_set.contains(&asset1));
-        assert!(finalization_set.contains(&asset3));
-    }
-}
